Small edits and making sure that my own login cannot be deleted.

+++ app.py
@@ -71,6 +71,12 @@ def get_settings_questions():
         settings1.append(setting[0])
     return questions1, settings1
 
+
+def get_authority(username):
+    sql = f"SELECT authority_lvl FROM admin_login WHERE username = '{username}';"
+    authority = query_db(sql)[0]
+    return authority
+
 @app.teardown_appcontext
 def close_connection(exception):
     db = getattr(g, '_database', None)
@@ -263,8 +269,8 @@ def new_user():
             hashed_pw = generate_password_hash(password)
             with sqlite3.connect(DATABASE) as db:
                 cursor = db.cursor()
-                sql = """INSERT INTO admin_login (username, encrypted_pw, authority_lvl)
-                VALUES ('""" + username + "', '" + hashed_pw + "', 1);"
+                sql = """INSERT INTO admin_login (username, encrypted_pw)
+                VALUES ('""" + username + "', '" + hashed_pw + "');"
                 cursor.execute(sql)
             error_message = "New admin '" + username + "' created successfully"
     return render_template("admin.html", setting_names=data_dict.keys(), setting_data=json.dumps(data_dict), everything=json.dumps(everything_dict), error_message=error_message)
@@ -281,16 +287,19 @@ def remove_user():
         # Find existing usernames
         sql = 'SELECT username FROM admin_login WHERE username = "' + username + '";'
         overlap = query_db(sql)
-        if overlap:
-            with sqlite3.connect(DATABASE) as db:
-                cursor = db.cursor()
-                sql = """DELETE FROM admin_login
-                WHERE username = '""" + username + "';"
-                cursor.execute(sql)
-            error_message = "Admin '" + username + "' has been deleted"
+        if username != '23595':
+            if overlap:
+                with sqlite3.connect(DATABASE) as db:
+                    cursor = db.cursor()
+                    sql = """DELETE FROM admin_login
+                    WHERE username = '""" + username + "';"
+                    cursor.execute(sql)
+                error_message = "Admin '" + username + "' has been deleted"
+            else:
+                error_message = 'Username not found. Check spelling and capitalisation.'
+                # Returns an error message if it cannot find the username
         else:
-            error_message = 'Username not found. Check spelling and capitalisation.'
-            # Returns an error message if it cannot find the username
+            error_message = 'This user cannot be deleted.'
     return render_template("admin.html", setting_names=data_dict.keys(), setting_data=json.dumps(data_dict), everything=json.dumps(everything_dict), error_message=error_message)
